+++ KitchenConnect/screens/shared/loadingScreen.js
@@ -1,7 +1,7 @@
 import { StyleSheet, ActivityIndicator, View } from "react-native";
 import React from "react";
 
-const LoadinScreen = () => {
+const LoadingScreen = () => {
   return (
     <View style={styles.container}>
       <ActivityIndicator size="large" color="#ffa500" />
@@ -9,7 +9,7 @@ const LoadinScreen = () => {
   );
 };
 
-export default LoadinScreen;
+export default LoadingScreen;
 
 const styles = StyleSheet.create({
   container: { flex: 1, justifyContent: "center", alignItems: "center" },
