+++ KitchenConnect/screens/shared/welcomeScreen.js
