+++ KitchenConnect/styles/shared/activeScreen.js
