+++ KitchenConnect/styles/shared/authAd.js
